p4i integration Integrate 205994 from 4.1 to main: Fixes:    Potential crash when a widget is created before the widget mapper Task:     89253 RevBy:    Lars Details:  If a widget happened to be created before the widget mapper, we could end up           never deleting the widgets window on destruction, which may cause windows so           send Window messages to a procedure which does not exist in memory anymore,           and crash. This would be because the win ID would not be stored in the widgets           internal structure, if the widget mapper wasn't already created.
[git-p4: depot-paths = "//depot/qt/main/": change = 206355]

+++ src/gui/kernel/qwidget.cpp
@@ -1012,16 +1012,14 @@ int QWidgetPrivate::maxInstances = 0;     // Maximum number of widget instances
 void QWidgetPrivate::setWinId(WId id)                // set widget identifier
 {
     Q_Q(QWidget);
-    if (!mapper)                                // mapper destroyed
-        return;
-    if (data.winid)
+    if (mapper && data.winid)
         mapper->remove(data.winid);
 
     data.winid = id;
 #if defined(Q_WS_X11)
     hd = id; // X11: hd == ident
 #endif
-    if (id)
+    if (mapper && id)
         mapper->insert(data.winid, q);
 }
